VERSION 3.0.0-b021
A small but important fix: make sure that the JMXRegistrationManager
does not register the root in the following case:
- The root parent is registered with JMX before the MOM is created
- Registrations are suspended in the MOM
In this case, the root needs to be added to the deferred registrations
list, and the root is marked as suspended.  This prevents the
rpListener from registering the root even though the suspendCount > 0.

This causes a number of problems in the ORB, so this fix is important.